pipelines/requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
httpx>=0.27.0
google-auth>=2.20.0
